data/color_product/requirements.txt
# 이미지 컬러 추출 스크립트 필요 패키지

openai>=1.0.0
python-dotenv>=1.0.0
